DEV: adds robots.txt file (#3)
This PR adds a robots.txt file to serve at `/robots.txt`

We don't have one right now, so we show a 404 for that path. This PR fixes that. 

The content is basic for now. It disallows the bots listed on 

https://robotstxt.com/ai

While allowing other bots to crawl the site.

Note that respecting robots.txt is not something we control. It's up to the bot. 

That said, having one is better than a 404 page at least as far as Googlebot is concerned.

---------

Co-authored-by: Erlend Sogge Heggen <e.soghe@gmail.com>