PERF: load youtube video iframe after a click (#20)
Right now we have a youtube video on the landing page loaded via an iframe. 

This is fine, but youtube video iframes load a lot of scripts. 

![before](https://github.com/user-attachments/assets/1a281a54-8c96-4d26-b0ed-71e078c30359)

What this PR does is that it creates a new reusable component that you can pass the `videoId` and `title` to and it would then

1. get the thumbnail for the video (which is also lazy-loaded)
2. show it with a "play" button on top
3. when it's clicked, it will swap the thumbnail with the iframe

This means that we defer a ton of javascript (which comes from the youtube iframe) until the user shows clear intent.

This change cuts down landing page resource size from 4.4mb 655kb (before compression)

![after](https://github.com/user-attachments/assets/5be05d76-efce-4c63-8a10-e081c5dc751f)

There are 0 visual changes beyond that the video now shows up as a thumbnail until you click on it.