DEV: manually create the Github star button (#21)
Continuing from #17 

Right now, we have a Github button that links to the Weird repository and shows the star count.

It's currently implemented with a script that Github provides. This is fine, but we're trying to cut down on third party code since this is a static website.

This PR recreates the button without the Github script. It still looks and acts the same.

Before:

![image](https://github.com/user-attachments/assets/a54b78bf-e677-48f1-bd22-b5856164f023)


After:

![image](https://github.com/user-attachments/assets/b89bdf35-a6fc-48c5-b4d2-a690d1f091c4)

This PR moves us from 655kb to 639kb (before compression)

There's still a TODO here to get the star count at build-time but that will be in a separate PR.